Update Style.md
The example given is confusing. It is best to have styles that do not have conflicting properties for the learner to get the concept first. The use of 'color' attribute in both makes red not render and hence the output is misleading. I propose to maintainone style for controlling the size and the other to control the size.